backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
torch==2.1.0
torchvision==0.16.0
bcrypt==4.1.1
PyJWT==2.8.0
email-validator==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
